Product Controller, docs, CQRS 적용

+++ semicolon/src/main/java/dukku/semicolon/boundedContext/product/app/usecase/product/DeleteProductUseCase.java
@@ -1,8 +1,10 @@
 package dukku.semicolon.boundedContext.product.app.usecase.product;
 
+import dukku.common.global.eventPublisher.EventPublisher;
 import dukku.semicolon.boundedContext.product.app.support.ProductSupport;
 import dukku.semicolon.boundedContext.product.entity.Product;
 import dukku.semicolon.shared.product.event.ProductDeletedEvent;
+import dukku.semicolon.shared.product.exception.ProductUnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
@@ -14,13 +16,18 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class DeleteProductUseCase {
     private final ProductSupport productSupport;
-    private final ApplicationEventPublisher eventPublisher;
+    private final EventPublisher eventPublisher;
 
     @Transactional
     public void execute(UUID productUuid, UUID sellerUuid) {
         Product product = productSupport.getProduct(productUuid, sellerUuid);
+
+        if (!product.getSellerUuid().equals(sellerUuid)) {
+            throw new ProductUnauthorizedException();
+        }
+
         product.delete();
 
-        eventPublisher.publishEvent(new ProductDeletedEvent(product));
+        eventPublisher.publish(new ProductDeletedEvent(product));
     }
 }
